+++ NEWS.md
@@ -1,7 +1,6 @@
 #### DoddLabMetID 0.1.0
 * Added a `NEWS.md` file to track changes to the package.
 
-
 #### DoddLabMetID 0.1.1
 * Modify some functions to automatically modify the peak table
 * Add parameter set for calling
@@ -11,3 +10,23 @@
 #### DoddLabMetID 0.1.11
 * Fix the errors in the merge_one_mode function
 * Update README file
+
+#### DoddLabMetID 0.1.12
+* Modify read_ms2 function to extract ms2_purity for mzml/mzxml data
+* add the ms2_purify output
+
+#### DoddLabMetID 0.1.13
+* add peptide database & modify related functions
+* modify initialize_annotation_parameter_class function to assign ms2_file in the ms2 folder
+
+#### DoddLabMetID 0.1.14
+* add export of combine_ms1_ms2 function
+
+#### DoddLabMetID 0.1.15
+* add gnps_bile_acid, gnps_acyl_amides, gnps_acyl_esters, all_public options in database
+* add parameter export
+* add object slot in the AnnotationParameterClass
+
+#### DoddLabMetID 0.1.16
+* adjust the merge_one_modes function
+* add merge and export ms2 spec for manual check